# Reland 3rd try [finishing colesbury's PR 100642] Guard on nn.Module dicts and type