--- a/benchmarks/dynamo/ci_expected_accuracy/inductor_torchbench_dynamic_inference.csv
+++ b/benchmarks/dynamo/ci_expected_accuracy/inductor_torchbench_dynamic_inference.csv
@@ -9,8 +9,7 @@ basic_gnn_edgecnn,pass,0
 basic_gnn_gcn,pass,6
 basic_gnn_gin,pass,0
 basic_gnn_sage,pass,0
-clip,pass_due_to_skip,0
-cm3leon_generate,pass,5
+cm3leon_generate,pass,7
 dcgan,pass,0
 dlrm,pass,0
 doctr_det_predictor,pass,2