--- a/docs/source/torch.compiler_nn_module.rst
+++ b/docs/source/torch.compiler_nn_module.rst
@@ -1,5 +1,5 @@
-PyTorch 2.0 NNModule Support
-============================
+PyTorch 2.0 nn.Module Support
+=============================
 
 **Author**: `Will Constable <https://github.com/wconstab>`_
 
@@ -8,12 +8,9 @@ arbitrary python classes, with the intent of producing faster code by making ass
 
 This doc describes some of the tradeoffs or edge cases that come up due to this specialization.
 
-NNModule Hooks Support
-----------------------
-Previously, `torch.compile` had no support for hooks on nn.Modules, and if hooks were registered
-they would simply be ignored in the compiled program. Indeed many users do not
-use nn.Module hooks at all, or only use them for debug workflows, but there are valid use cases
-for composing nn.Module hooks with `torch.compile`.
+`nn.Module` Hooks Support
+-------------------------
+`torch.compile` now has partial support for forward and backward hooks on nn.Modules.
 
 Hooks that are orchestrated via nn.Module.__call__ implementation include `_forward_pre_hooks`,
 `forward_hooks`, `_backward_pre_hooks`, and `_backward_hooks`, and will be referred to as 'call hooks'.
@@ -25,11 +22,11 @@ unsupported by `torch.compile`.
 `nn.Module.__call__` Hooks Usage and limitations
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 By default, `torch.compile` will trace the contents of `nn.Module.__call__` which means it will encounter
-and run forward/pre-forward hooks.  If you install hooks before calling `torch.compile` and then do not remove
-or alter the hooks later, your use case should be supported by default.
+and run forward/pre-forward hooks. `torch.compile` installs guards that detect added and removed hooks,
+and will trigger a recompilation if the forward/pre-forward hooks change.
 
 Backward/Pre-backward hooks are generally also supported, with similar caveats: currently graph-breaks in dynamo
-occur when accessing backward_hooks dicts, which is probably avoiable with some work.  Graph-breaks also impact the
+occur when accessing backward_hooks dicts, which is probably avoidable with some work.  Graph-breaks also impact the
 timing of firing backward hooks, since graph-segments are run as autograd-functions which produce all their grads at
 the same time.  Assuming it were possible for dynamo to not graph-break on the presence of backward-hooks, we would
 still expect the backward hooks for a series of modules to all fire together after the whole compiled graph's backward
@@ -40,17 +37,6 @@ ran.
 by allowing them to be called opaquely in the dynamo graph instead of traced into by dynamo.  For such modules, hooks
 currently trigger a graph-break so that the affected modules run outside of dynamo.  Depending on the model, this could
 introduce a significant performance regression, and additional work is required to improve this support.
-
-**skip_nnmodule_hook_guards**
-By default, `torch._dynamo.config.skip_nnmodule_hook_guards` is set to True, meaning no guards will be installed
-on each nn.Module hook dictionary, improving runtime by reducing guard execution time, at the cost of not noticing
-if any hook dict is changed after compilation.
-
-If you want to be able to remove or modify hooks after compilation and have `torch.compile` react appropriately
-(by recompiling), then you need to set `skip_nnmodule_hook_guards=False` and expect a runtime penalty for the added
-guards.
-
-TODO: confirm if backward/pre_backward hooks are working or not and document accordingly
 
 state_dict Hooks
 ~~~~~~~~~~~~~~~~