--- a/test/dynamo/test_modules.py
+++ b/test/dynamo/test_modules.py
@@ -1247,6 +1247,7 @@ class NNModuleTests(torch._dynamo.test_case.TestCase):
         module_dict = torch.nn.ModuleDict({"cat": torch.nn.Conv2d(1, 1, 1)})
         pre = m(data)
         cnt.clear()
+        torch._dynamo.reset()
 
         with torch._dynamo.optimize(cnt, nopython=False):
             opt_pre = m(data)
@@ -1255,8 +1256,8 @@ class NNModuleTests(torch._dynamo.test_case.TestCase):
             out1 = m(data)
 
         out_post = m(data)
-        self.assertEqual(cnt.frame_count, 1)
-        self.assertEqual(cnt.op_count, 1)
+        self.assertEqual(cnt.frame_count, 2)
+        self.assertEqual(cnt.op_count, 2)
         self.assertTrue(torch._dynamo.testing.same(pre, opt_pre))
         self.assertTrue(torch._dynamo.testing.same(out1, out_post))
 
@@ -1730,7 +1731,112 @@ class OptimizedModuleTest(torch._dynamo.test_case.TestCase):
             )
         )
 
-    @patch.object(torch._dynamo.config, "skip_nnmodule_hook_guards", False)
+    def test_hooks_recompile(self):
+        # Modifying hooks should lead to a recompiation
+        class TestModule(torch.nn.Module):
+            def forward(self, x: torch.Tensor) -> torch.Tensor:
+                return 2 * x + 1
+
+        def compute_output_and_grad(m, x):
+            output = m(x)
+            output.sum().backward()
+            return x.grad
+
+        def forward_pre_hook(module: torch.nn.Module, inputs: Tuple[torch.Tensor]):
+            return (2 * inputs[0] + 1,)
+
+        def forward_hook(
+            module: torch.nn.Module, inputs: Tuple[torch.Tensor], output: torch.Tensor
+        ):
+            return 2 * output + 1
+
+        def backward_hook(module, grad_input, grad_output):
+            if len(grad_input) == 1:
+                return (grad_input[0] * 3,)
+            else:
+                return (grad_input[0] * 3, None)
+
+        def backward_pre_hook(module, grad_outputs):
+            return (grad_outputs[0] * 5,)
+
+        def run_test_case(hook_type, hook_func, expected_grad):
+            torch._dynamo.reset()
+            m = TestModule()
+            input = torch.ones(10, requires_grad=True)
+            cnt = torch._dynamo.testing.CompileCounter()
+            opt = torch._dynamo.optimize(cnt)(compute_output_and_grad)
+
+            grad1 = opt(m, input)
+            self.assertEqual(cnt.frame_count, 1)
+            self.assertEqual(grad1, torch.full_like(grad1, 2))
+
+            input.grad = None
+            handle = getattr(m, hook_type)(hook_func)
+            grad2 = opt(m, input)
+            frame_count2 = cnt.frame_count
+            # Some backward hooks lead to graph breaks so frame_count may be 2 or 3
+            self.assertGreaterEqual(frame_count2, 2)
+            self.assertEqual(grad2, torch.full_like(grad2, expected_grad))
+
+            # Running again should not recompile
+            opt(m, input)
+            self.assertEqual(cnt.frame_count, frame_count2)
+
+            # Removing handle should lead to original result
+            input.grad = None
+            handle.remove()
+            grad3 = opt(m, input)
+            self.assertEqual(grad1, grad3)
+
+        run_test_case("register_forward_pre_hook", forward_pre_hook, 4)
+        run_test_case("register_forward_hook", forward_hook, 4)
+        run_test_case("register_backward_hook", backward_hook, 6)
+        run_test_case("register_full_backward_hook", backward_hook, 6)
+        run_test_case("register_full_backward_pre_hook", backward_pre_hook, 10)
+
+    def test_unspecialized_nn_module(self):
+        # This test is little confusing because of combination of
+        # nn_module_guard and unspecialized nn module variable.
+
+        # The graph break in forward causes two graphs
+        # 1) The first graph has self.relu which introduces a nn_module_guard
+        # 2) The second graph first assumes self to be NNModuleVariable, but the
+        # restarts the analysis with self mapping to
+        # UnSpecializedNNModuleVariable, on witnessing self.a += 1.
+
+        # Now, when we run the compiled mod the first time, it changes the value
+        # of self.a. This is fine for the first run. But, when we run the
+        # compiled module again, the first graph recompiles. This is because
+        # self.a has changed, changing the ma_version_tag, causing
+        # nn_module_guard to fail.
+
+        # At this point, we might feel that this is doomed as we will always
+        # keep recompiling on the first graph. But, then Dynamo has already
+        # marked the self to be UnspecializedNNModuleVariable (because of self.a
+        # in the second graph), and therefore during the recompilation, we do
+        # not introduce any nn_module_guard. So, in all we have just one
+        # recompilation.
+        class Mock(torch.nn.Module):
+            def __init__(self):
+                super().__init__()
+                self.a = 5
+                self.relu = torch.nn.ReLU()
+
+            def forward(self, x):
+                z = self.relu(x)
+                torch._dynamo.graph_break()
+                self.a += 1
+                return z * self.a
+
+        mod = Mock()
+        cnt = torch._dynamo.testing.CompileCounter()
+        opt = torch.compile(mod, backend=cnt)
+
+        for _ in range(5):
+            opt(torch.randn(4))
+
+        self.assertEqual(cnt.frame_count, 4)
+
     def test_hooks_outer(self):
         class TestModule(torch.nn.Module):
             def forward(self, x: torch.Tensor) -> torch.Tensor:
@@ -1777,7 +1883,6 @@ class OptimizedModuleTest(torch._dynamo.test_case.TestCase):
             the eval_frame entrypoint to Module.__call__?
         """
 
-    @patch.object(torch._dynamo.config, "skip_nnmodule_hook_guards", False)
     def test_hooks_inner(self):
         class TestModule(torch.nn.Module):
             def forward(self, x: torch.Tensor) -> torch.Tensor:
@@ -1822,7 +1927,7 @@ class OptimizedModuleTest(torch._dynamo.test_case.TestCase):
         handle.remove()
         self.assertEqual(compiled_func(inp), outer_func(inp))
         self.assertEqual(compiled_func(inp).item(), 7)
-        self.assertTrue("forward_hooks.keys" in failure_reason)
+        self.assertTrue("__nn_module_guard" in failure_reason)
         self.assertEqual(cc.frame_count, 1 + 1)
         self.assertEqual(cc.op_count, 6 + 4)
 
@@ -1842,55 +1947,7 @@ class OptimizedModuleTest(torch._dynamo.test_case.TestCase):
         m._forward_hooks[handle.id] = new_forward_hook
         self.assertEqual(compiled_func(inp), outer_func(inp))
         self.assertEqual(compiled_func(inp).item(), 16)
-        self.assertTrue("___check_obj_id(L['m']._forward_hooks" in failure_reason)
-
-    @patch.object(torch._dynamo.config, "skip_nnmodule_hook_guards", True)
-    def test_hooks_skip_guards(self):
-        class TestModule(torch.nn.Module):
-            def forward(self, x: torch.Tensor) -> torch.Tensor:
-                return 2 * x + 1
-
-        m = TestModule()
-
-        def forward_hook(
-            module: torch.nn.Module, inputs: Tuple[torch.Tensor], output: torch.Tensor
-        ) -> torch.Tensor:
-            return 2 * output + 1
-
-        handle = m.register_forward_hook(forward_hook)
-
-        def outer_func(tensor):
-            x = tensor * 2 + 1
-            y = m(x)
-            return y
-
-        inp = torch.tensor(1.0, requires_grad=True)
-
-        failure_reason = None
-
-        def guard_fail_fn(failure):
-            nonlocal failure_reason
-            failure_reason = failure[0]
-
-        cc = torch._dynamo.testing.CompileCounterWithBackend("aot_eager")
-        compiled_func = torch._dynamo.optimize(
-            guard_fail_fn=guard_fail_fn,
-            backend=cc,
-        )(outer_func)
-
-        m = TestModule()
-        handle = m.register_forward_hook(forward_hook)
-        failure_reason = None
-        self.assertEqual(compiled_func(inp), outer_func(inp))
-        self.assertEqual(compiled_func(inp).item(), 15)
-        self.assertEqual(cc.frame_count, 1)
-        self.assertEqual(cc.op_count, 6)
-
-        # if we remove the hook, dynamo shouldn't notice
-        handle.remove()
-        self.assertNotEqual(compiled_func(inp), outer_func(inp))
-        self.assertEqual(compiled_func(inp).item(), 15)
-        self.assertEqual(cc.frame_count, 1)
+        self.assertTrue("__nn_module_guard" in failure_reason)
 
     def _forward_hook_test_helper(self, model):
         forward_handles = {}