--- a/torch/_dynamo/config.py
+++ b/torch/_dynamo/config.py
@@ -219,6 +219,8 @@ skip_fsdp_guards = True
 # Make dynamo skip guarding on hooks on nn modules
 # Note: unsafe: if your model actually has hooks and you remove them, or doesn't and  you add them,
 # dynamo will not notice and will execute whichever version you first compiled.
+# TODO(janimesh) - Remove as once internal is not reliant on this flag. This
+# flag has no meaning now after nn_module_guard is introduced.
 skip_nnmodule_hook_guards = True
 
 # If True, raises exception if TorchDynamo is called with a context manager