@Valentine233 2023-10-11
It is found that this PR caused the regression of `hf_T5_generate` 2023-09-22 vs 2023-09-15 nightly release, because of the added guards on `nn.Module`.

name | batch_size_new | speed_up_new | inductor_new | eager_new | compilation_latency_new | batch_size_old | speed_up_old | inductor_old | eager_old | compilation_latency_old | Ratio Speedup(New/old) | Eager Ratio(old/new) | Inductor Ratio(old/new) | Compilation_latency_Ratio(old/new)
-- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | --
hf_T5_generate | 1 | 1.098084 | 1.306571603 | 1.4347253721086521 | 328.140796 | 1 | 1.311963 | 1.105396172 | 1.4502388780056361 | 161.974814 | 0.84 | 1.01 | 0.85 | 0.49