@blzheng 2023-10-11
Also observed that this PR caused the regression of `lennard_jones`. Before this pr, the speedup is **1.068x**, while after this pr, the speedup is **0.951x**.